Update index.html
To play a sound when the "Biodata" button is clicked, you'll need to embed an audio player in your HTML and trigger it using JavaScript. Here's how you can modify your code to play a sound from your Google Drive link.

Google Drive audio links need to be shared publicly and converted into a downloadable format (direct link) for proper use.
Since the provided link isn't a direct audio file link, it needs to be adjusted. For the purpose of this solution, we'll assume that the audio file can be played directly by embedding the correct format URL.
Here's how to modify your code:

Steps:
Embed an audio player in the HTML.
Modify the toggleBiodata function to play the sound when clicked.